+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
-* `signing.sign_get_url()` is deprecated and will be removed in a future release.
+* `signing.sign_get_url()` is deprecated and will be removed in a future
+  release.
+* `LaterPayClient.get_access()` is deprecated and will be removed in a future
+  release. Consider using `LaterPayClient.get_access_data()` instead.
 
 
 ## 4.1.0